moving slider config to other file
